refactoring Images Slider
