+++ pages/projects/[projectId]/carousel-with-stepper/slider/styles.tsx
@@ -1,0 +1,7 @@
+export const sliderStyles = {
+    height: "500px",
+    width: "700px",
+    display: "block",
+    maxWidth: "100%",
+    overflow: "hidden",
+};
